Fixed the (hopefully) last implicit cast warnings

+++ OpenNeighborhood/src/Core/Window.cpp
@@ -61,10 +61,10 @@ void Window::Init(const WindowProps& props)
 	glfwSetWindowSizeCallback(m_Window, [](GLFWwindow* window, int width, int height)
 		{
 			WindowData& windowData = *(WindowData*)glfwGetWindowUserPointer(window);
-			windowData.Width = width;
-			windowData.Height = height;
+			windowData.Width = (float)width;
+			windowData.Height = (float)height;
 
-			WindowResizeEvent event(width, height);
+			WindowResizeEvent event((float)width, (float)height);
 			windowData.EventCallback(event);
 		});
